Added form functionality and updated database structure
This added new template allowing a user to select an option for a question and a view for
recording that response in the database. Also renamed 'response' to 'option' and added a new seperate
response table. The option table records options for a question (using question as an fk) while the response table
records a given response was entered (using an option as an fk).