Began adding user authentication and restructured forms models.
Added a few templates in the main file tempaltes folder to handle user login and some
password reset features, though this is very much still a WIP. Also restructured to models
from dummy tables to a first crack at how the database would be structured in production. The
forms app is not currently functional, as views and tempaltes need to be udpated for the
new models.