Began working on Views and Admin for Forms.
I began working on some new views for forms. I pretty quickly realized that I should
also create admin functionality for forms as well for ease of testing the UI of the views.
Views do not currently work, and Admin forms are in progress. Currently trying to work on
the option message in the jquery for the question admin form.