Added the ability to view and edit responses.
The biggest update with this batch of changes is that users with permissions can now edit
individual responses. This change also included a pretty significant restructuring of the urls.
Filling out a new form is now considered a 'new response' and follows the response url instead of
being the default for form/id, which will be replaced with a detail view. Also tried to clean
up some variable names and file names to make things a bit clearer.