Updated Create/edit form view.
The create and edit form view is now much more useful. Users
can now move questions around, insert questions positionally,
and edit/create questions from the edit form menu (if no question is selected,
it will link the user to a create question view). A big challenge
with this is that to create or edit a question, you would
have to refresh the edit form page for it to take effect, which
would remove all the current edits.