Reworked permissions and added user profile for organizational affiliations.
The main update is new functionality that assigns an organization to each user via
a new UserProfile model linked to each user. As of now, users are limited in what forms
they can view based on their organization. Also I reworked some typos in
permissions (it should be add/change, not create/edit). Also, I realized that the
conditional logic was broken since the html structure of django fields differs by type.
I reworked it to make it a little more robust.