Added file upload/template creation functionality to forms.
In the form detail page, a user can now create a csv template with respondent details
and form questions, input data into the tempalte, and then upload the template to create
responess in bulk. This method is still pretty janky (there's bascially
no data validation and no checks for duplicate uploads, so it would theoretically be very
easy to totally screw up the db with an upload) but the basic functionality works.
Users can input answers to multiple choice questions by either typing the option id or by
typing the option text (the program will try to match it), but this must be matched exactly,
including case.