Added dashboard tab and question chart.
This commit includes a new dashboard tab that allows users to access the prototypical
chart feature to view forms data as charts. So far, the only charts are one that shows
forms by responses and another that allows a user to select a question and then
view the number of responses for each answer.