Dashboard Creator, minor restyling, and cleaned up form logic.
Added a 'dashboard creator', which currently only allows users to add a chart and select
a question to view data about, but in the future I anticipate adding the ability to
create different chart types. This kicked off a bit of an effort to restyle some pages,
which also led me to find a couple bugs in the javascripts that controlled
form creation and response creation.

Form creation is nominally the same, but the code looks a bit cleaner since button creation
is now done in a function. Also added in verification to ensure that fields that appear
conditionally require a user to enter a response. All the javascript associated with the project
is low key kinda trash, but we're making it slightly better (I hope) one commit at a time.