Updated User Access structure and added validation to creating/updating questions.
The permission checks in the tempaltes were a bit of a mess, so I decided to create
access_levels for each user and house them in the UserProfile model. These are currently
data collector, supervisor, manager, and admin. Over the next few commits,
ideally most of the templates will be updated to fir this new structure.

Also, I added basic validaiton and styling to the create/update question view,
as well as finally creating a view question screen. Questions are only editable by admins,
since ideally questions are standardized and assigned to all organizations (like objectives) and
should be the same across all orgnaizations so that comparing data is easier.