Began Updating Conditional Form Logic
I'm beginning an overhaul of conditional form logic, since before it was pretty trash
and didn't work most of the time. This new system is much more robust, giving users the ability
to select options, hide options for the conditional question, and combine logic.

Currently I've only built out the models and the form creation aspects. Next is
actually making the logic work when filling out the forms. Also I may rethink the operators part
so that users can combine logic, but I'm not sure there are enough use cases to make this worth it.
We'll see.