Added more functionality to conditional logic.
Fixed some bugs with conditional logic and added a new feature where for text inputs,
a user can select comparison for equal to, contains, or does not contain and for numbers
a user can select comparison for equal to, greater than, or less than. Also added checking logic
to make sure conditional fields are filled in.