Added new field to option allowing for a select none.
There is a now a special field (currently only works for multi-select) that allows for a user
to select (none of the above). This will deselect any other selected inputs and record to response as
null, but will fullfill the requirement for the question being answered.

I also slightly reworked the logic so that questions won't show if their logic is
anything selected but none of the above is selected.

There's currently the option to select all, but I have't implemented this yet. Also I'm debating whether
this should be extended to single select, since it might not be necessary, since the main appeal of
this feature is dynamic field checking.