Working on redoing form creation.
Form creation is a total mess, both in terms of the script and in the user experience. I'm
going to take some time to rework it so that the scripts are more readable and there is more communication
to the user about what's going on, especially with regards to logic. Its a WIP.