Improved Chart creator.
The chart creator is lowkey kind of epic. We added filters and axis sorting and the ability
to toggle showing a legend or just the raw count of unique responses that are not none or no.

It probably doesn't work that well since it hasn't been tested that thoroughly, but so far so
good, so ¯\_(ツ)_/¯.

Next up the plan is to add the capability for a person to add targets to the graph.