Slightly reworked code and added targets.
Made a few minor edits to make code more readable and also added the ability
for users to view targets (which I think works after some testing).

Filters work with targets, but currently if a filter is unable to work with a target,
it just sends a console warn. Definitely need to add some better styling and user feedback
to this system, but its decent for now.