Added Messages, redid response recording, and too much.
Added a messaging feature, pretty simple, you can send messages and reply to them. There is also
a mark as complete and read functionality.

I refactored responses to updating and creating use the same function which is hopefuly more robust now.
Also slightly reworked the response form form to accompany this.
Continued testing and made a few bug fixes and improvements here and there as well.